ISSUE #4321 - FIxed circular dependencies between api files
