+++ frontend/package.json
@@ -29,7 +29,8 @@
 		"extract": "node ./internals/intl/extract.js",
 		"compile": "node ./internals/intl/compile.js",
 		"storybook": "start-storybook -p 6006",
-		"build-storybook": "build-storybook -o ../public/storybook"
+		"build-storybook": "build-storybook -o ../public/storybook",
+		"circular:test": "yarn dpdm src/main.tsx --exclude node_modules/ --warning false --tree false"
 	},
 	"keywords": [
 		"3drepo",
