+++ frontend/src/v4/services/api/index.ts
@@ -16,7 +16,7 @@
  */
 
 import { API } from './default';
-export { getAPIUrl, getResponseCode, setSocketIdHeader } from './default';
+// export { getAPIUrl, getResponseCode, setSocketIdHeader } from './default';
 
 export * from './auth';
 export * from './users';
