+++ frontend/src/v4/services/api/invitations.ts
@@ -14,8 +14,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import api from './';
+import { API as api } from './default';
 
 /**
  * Get invitations
