+++ frontend/src/v4/services/api/tree.ts
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import api from './';
+import { API as api } from './default';
 
 /**
  * Get full model/federation tree
