+++ frontend/src/v4/services/export.ts
@@ -15,15 +15,18 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as API from '../services/api';
+import { downloadJSON } from "./api";
+import { getAPIUrl } from "./api/default";
+
+// import * as API from '../services/api';
 
 export const exportBCF = (teamspace, model, issueNumbers) => {
-	const exportUrl = API.getAPIUrl(`${teamspace}/${model}/issues.bcfzip?numbers=${issueNumbers}`);
+	const exportUrl = getAPIUrl(`${teamspace}/${model}/issues.bcfzip?numbers=${issueNumbers}`);
 	window.open(exportUrl, '_blank', 'noopener');
 };
 
 const handlePrint = (dataType) => (teamspace, model, dataNumbers) => {
-	const printUrl = API.getAPIUrl(`${teamspace}/${model}/${dataType}.html?numbers=${dataNumbers}`);
+	const printUrl = getAPIUrl(`${teamspace}/${model}/${dataType}.html?numbers=${dataNumbers}`);
 	window.open(printUrl, '_blank', 'noopener');
 };
 
@@ -33,7 +36,7 @@ export const printRisks = handlePrint('risks');
 
 const handleExportToJSON = (dataType) => (teamspace, model, dataNumbers) => {
 	const endpoint = `${teamspace}/${model}/${dataType}?numbers=${dataNumbers}&convertCoords=1`;
-	return API.downloadJSON(dataType, model, endpoint);
+	return downloadJSON(dataType, model, endpoint);
 };
 
 export const exportIssuesToJSON = handleExportToJSON('issues');
